requirements.txt
boto3==1.28.75
botocore==1.31.75
jmespath==1.0.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
s3transfer==0.7.0
six==1.16.0
urllib3==2.0.7
